std.ChildProcess: Fix retry on FileNotFound when the ext is in the command
For example, if the command is specified as `something.exe`, the retry will now try:

```
C:\some\path\something.exe
C:\some\path\something.exe.COM
C:\some\path\something.exe.EXE
C:\some\path\something.exe.BAT
... etc ...
```

whereas before it would only try the versions with an added extension from `PATHEXT`, which would cause the retry to fail. Note that this is only the case when an env map is not provided to `exec`, if an env map with a `PATH` that has the command is given to `exec`, the first `windowsCreateProcess` will succeed and the retry branch won't be executed.